requirements.txt
# For the GUI
customtkinter
requests

# For the builder
pyinstaller

# For the payload itself
psutil
requests
pyautogui
pillow
pycaw
pywin32